feat: Signin and login page
This commit introduces the ability to sign in and login to the page.
The user can enter the email, username and password to signin and username and password to login.
The page will show error message if the user enter wrong user name or password or there is a server error.